ported HeliosMacroBoard to plugin
using private builds of dependencies as submodules
publishing macro board plugin as zip for alpha testing
adjusted build configurations to include macro board
cleanup and factoring into files
fixed binding definitions for macro buttons in ProfileEditor
    this allows the default binding of TriggerValue to work correctly
now works correctly if limited to a single device
