--- Helios/Interfaces/OpenMacroBoard/MacroBoardInterfaceEditor.xaml
@@ -1,44 +1,0 @@
-﻿<HeliosSdk:HeliosInterfaceEditor x:Class="GadrocsWorkshop.Helios.Interfaces.OpenMacroBoard.MacroBoardInterfaceEditor"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:System="clr-namespace:System;assembly=mscorlib"
-             xmlns:local="clr-namespace:GadrocsWorkshop.Helios.Interfaces.OpenMacroBoard"
-             xmlns:StreamDeckAlias="clr-namespace:GadrocsWorkshop.Helios.Interfaces.OpenMacroBoard"
-             xmlns:HeliosSdk="clr-namespace:GadrocsWorkshop.Helios.Windows.Controls"
-             DataContext="{Binding RelativeSource={RelativeSource Self}}"
-             Background="{StaticResource ToolBackground}" Padding="5" Height="Auto" Width="Auto">
-    <HeliosSdk:HeliosInterfaceEditor.Resources>
-        <ObjectDataProvider x:Key="streamDeckModels" MethodName="GetValues" ObjectType="{x:Type System:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="StreamDeckAlias:MacroBoardModel"/>
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <local:MacroBoardModelConverter x:Key="modelConverter"/>
-        <local:MacroBoardModelValuesConverter x:Key="modelValuesConverter"/>
-    </HeliosSdk:HeliosInterfaceEditor.Resources>
-    <StackPanel>
-        <GroupBox Header="Profile-Specific Settings">
-            <Grid>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="Auto" />
-                    <ColumnDefinition Width="*" />
-                </Grid.ColumnDefinitions>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-                <Label Grid.Column="0" Grid.Row="2" HorizontalAlignment="Right">Brightness</Label>
-                <TextBox Grid.Column="1" Grid.Row="2" Margin="0,3,0,0" HorizontalAlignment="Left" Width="40" MaxWidth="40" Text="{Binding Path=Interface.Brightness}" PreviewTextInput="ValidateBrightnessTextBox"/>
-                <TextBlock Grid.Column="1" Grid.Row="3" Style="{StaticResource Documentation}" Text="Brightness of the Stream Deck in percent."/>
-            </Grid>
-        </GroupBox>
-        <GroupBox Header="Stream Deck Button Configuration">
-            <StackPanel Orientation="Vertical">
-                <TextBlock Style="{StaticResource Documentation}" Text="To edit a button, right click it and select &quot;Properties&quot; in the context menu."/>
-                <local:MacroBoardLayoutEditor></local:MacroBoardLayoutEditor>
-            </StackPanel>
-        </GroupBox>
-    </StackPanel>
-</HeliosSdk:HeliosInterfaceEditor>
