+++ Helios/packages.config
@@ -9,10 +9,6 @@
   <package id="Newtonsoft.Json.Schema" version="3.0.13" targetFramework="net472" />
   <package id="NLog" version="4.7.0" targetFramework="net472" />
   <package id="NLua" version="1.4.31" targetFramework="net472" />
-  <package id="OpenMacroBoard.SDK" version="3.0.0" targetFramework="net472" />
-  <package id="OpenMacroBoard.VirtualBoard" version="4.1.0" targetFramework="net472" />
   <package id="SharpDX" version="4.2.0" targetFramework="net45" />
   <package id="SharpDX.DirectInput" version="4.2.0" targetFramework="net45" />
-  <package id="StreamDeckSharp" version="3.0.0" targetFramework="net472" />
-  <package id="System.Drawing.Common" version="4.5.1" targetFramework="net472" />
 </packages>
