+++ HeliosMacroBoard/HeliosMacroBoard.csproj
@@ -1,0 +1,89 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <ProjectGuid>{0B1B0665-99D5-4990-994D-8F92A0E15299}</ProjectGuid>
+    <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
+    <TargetFrameworkProfile />
+  </PropertyGroup>
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+  <PropertyGroup>
+    <RootNamespace>GadrocsWorkshop.Helios.Interfaces.HeliosMacroBoard</RootNamespace>
+  </PropertyGroup>
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <ItemGroup>
+    <Reference Include="NLog, Version=4.0.0.0, Culture=neutral, PublicKeyToken=5120e14c03d0593c, processorArchitecture=MSIL">
+      <HintPath>..\packages\NLog.4.7.0\lib\net45\NLog.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
+    <Reference Include="System" />
+    <Reference Include="System.Configuration" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Drawing" />
+    <Reference Include="System.IO.Compression" />
+    <Reference Include="System.Runtime.Serialization" />
+    <Reference Include="System.ServiceModel" />
+    <Reference Include="System.Transactions" />
+    <Reference Include="System.Xaml" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="MacroBoardButton.cs" />
+    <Compile Include="MacroBoardButtonPropertiesDialog.xaml.cs">
+      <DependentUpon>MacroBoardButtonPropertiesDialog.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="MacroBoardInterface.cs" />
+    <Compile Include="MacroBoardInterfaceEditor.xaml.cs">
+      <DependentUpon>MacroBoardInterfaceEditor.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="MacroBoardLayoutEditor.xaml.cs">
+      <DependentUpon>MacroBoardLayoutEditor.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="MacroBoardModel.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StreamDeckInterface.cs" />
+    <Compile Include="StreamDeckXLInterface.cs" />
+    <Compile Include="VirtualMacroboardInterface.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Page Include="MacroBoardButtonPropertiesDialog.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="MacroBoardInterfaceEditor.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="MacroBoardLayoutEditor.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Thirdparty\OpenMacroBoard.SDK\src\OpenMacroBoard.SDK\OpenMacroBoard.SDK.csproj">
+      <Project>{3e28549c-c22e-4545-a13d-32981a39ae91}</Project>
+      <Name>OpenMacroBoard.SDK</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Thirdparty\OpenMacroBoard.SDK\src\OpenMacroBoard.VirtualBoard\OpenMacroBoard.VirtualBoard.csproj">
+      <Project>{94715b35-6f71-403b-84d1-9839c017ced6}</Project>
+      <Name>OpenMacroBoard.VirtualBoard</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Thirdparty\StreamDeckSharp\src\StreamDeckSharp\StreamDeckSharp.csproj">
+      <Project>{7becb623-5eb3-44e4-b702-58ec3777f284}</Project>
+      <Name>StreamDeckSharp</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <Import Project="..\HeliosPlugin.targets" />
+</Project>
