+++ HeliosMacroBoard/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NLog" version="4.7.0" targetFramework="net472" />
+  <package id="System.Drawing.Common" version="4.5.1" targetFramework="net472" />
+</packages>
